Fixes bug with initially-selected button inside ViewGroup.
In the case where a `RadioButton` is defined with `android:checked="true"`, and that radio is inside another `ViewGroup` inside the `RadioGroupPlus`, the initially-checked state is never properly handled. When we try to check anything else, it doesn't realize that there was a checked view, so it doesn't get unchecked. Further, since the initial view is internally already in a checked state, tapping on it will never broadcast that it got checked, so no state changes happen.

This PR simply adds a check when traversing the tree to catch that case.